Move tabAnchors to use it in many components
