+++ src/app/pages/ValidatorDetailsPage/DebondingDelegationsCard.tsx
@@ -9,7 +9,7 @@ import { CardEmptyState } from '../../components/CardEmptyState'
 import { useSearchParamsPagination } from '../../components/Table/useSearchParamsPagination'
 import { Delegations } from '../../components/Delegations'
 import { ValidatorDetailsContext } from './hooks'
-import { debondingContainerId } from './tabAnchors'
+import { debondingContainerId } from '../../utils/tabAnchors'
 import { SimplePaginationEngine } from 'app/components/Table/PaginationEngine'
 import { LinkableCardLayout } from 'app/components/LinkableCardLayout'
