+++ src/app/pages/ValidatorDetailsPage/index.tsx
@@ -26,7 +26,7 @@ import { SignedBlocks } from './SignedBlocks'
 import { StakingTrend } from './StakingTrend'
 import { ProposedBlocks } from './ProposedBlocks'
 import { ValidatorDetailsContext } from './hooks'
-import { debondingContainerId, delegatorsContainerId } from './tabAnchors'
+import { debondingContainerId, delegatorsContainerId } from '../../utils/tabAnchors'
 import { ValidatorStatusBadge } from './ValidatorStatusBadge'
 import { eventsContainerId } from './../../pages/ConsensusAccountDetailsPage/ConsensusAccountEventsCard'
 import { PercentageValue } from '../../components/PercentageValue'
